Move spz parser to scripts/esm/parsers
